fix: rename modal-window-store to follow one case style + modify the store itself
